feat: add AccessLogRepo for managing access logs and related queries
